Render something

+++ RayTracerRedone/tracer/scene/ViewPlane.h
@@ -4,8 +4,8 @@
 class ViewPlane
 {
 public:
-	const float wsize;
-	const float hsize;
+	const uint32_t wsize;
+	const uint32_t hsize;
 	const float pixelsize;	
 
 	ViewPlane(uint32_t _hsize, uint32_t _vsize, float _pixelsize) :wsize(_hsize), hsize(_vsize), pixelsize(_pixelsize) {}
